Refactor imports, add index.ts to each folder
