+++ src/entities/user/index.ts
@@ -1,0 +1,5 @@
+export * from './user.controller';
+export * from './user.interface';
+export * from './user.route';
+export * from './user.service';
+export * from './user.validation';
